examples/gpu-accel/requirements.txt
tensorflow-gpu==2.4.0
torch==1.8.1+cu111
torchvision==0.9.1+cu111
numpy==1.19.5
pandas==1.3.0
scikit-learn==0.24.2
matplotlib==3.4.2
tensorboard==2.4.0
horovod==0.22.1
opencv-python==4.5.1.48 